careermatch/requirements.txt
datapizza-ai
python-dotenv
pydantic
requests
beautifulsoup4
fastapi
uvicorn
jinja2
python-multipart
pypdf